Restructure cli app into an agent
The wiregarden client should really be an agent that mediates between
logical devices defined by the wiregarden API and local network
interfaces.

The lifecycle of the network interface is modeled as a state machine
with transitions driven by joining, refreshing, or departing the
wiregarden network. Interface state and configuration is managed in a
SQLite database.

This nails down some edge cases with the simple JSON config rewriting in
the prior client, with proper modeling of current, pending, and past
applied states. Not only is this more observable, this design allows
operating wiregarden as a background process that automatically
discovers and reacts to topology and configuration changes in the
networks that it is joined to. It also allows for automatically retrying
failed operations.