Fix router deploy, more robust agent with backoff refresh.
Refresh on a decaying schedule from network changes, on the premise that
if a network has a change, it's likely that there may be others in the
near future, decreasing with probability over time.

Fix router deploy and provisioning of internal wiregarden deployment
network.

Validate endpoint in client and server better.

SKIP_BUILD option to speed up deploys.